Import library after registering GResources
When moving to GTK widget templates in the following commit,  we need to
make sure the library's GResource is registered before importing modules
that define a widget class with a template.

Therefore, in PresenterLoader, we delay importing Presenter and
ReaderPresenter, respectively, until just before creating the presenter.
We also have to register the GResource early in some tests.

[endlessm/eos-sdk#3281]