Card interface
This is a first iteration of a Card interface based on what the current
cards do. It takes care of fading its widget in if the construct property
"fade-in" is true.

The "fade-in" property is temporary, in order to reduce the amount of
code churn; in a later commit this should be replaced by a fade_in()
method, since the property is not actually used except during show_all().

The Card interface also has a "title-capitalization" property which will
specify how to format the title. This is also somewhat of a workaround
since eventually we will want to do this in CSS with the "text-transform"
property, but that is not implemented in GTK right now.

In the following commits we will make it so this interface is backed by a
card model, and convert the existing cards to use the interface.

[endlessm/eos-sdk#3281]