Adapt ContentObjectModel to Card interface UI
This is an adapter for ContentObjectModel and its friends, representing
records in the database, to provide data for widgets implementing the
Card interface.

For each property of ContentObjectModel that might be displayed, the Card
interface expects the implementing widget to have a member with a
particular name (such as "title_label" for the "ContentObjectModel.title"
property.) If this member is present, then the Card interface will
display the data in it; all members are optional though, so if a Card
implementation doesn't have a "title_label", that means that that card
prefers not to display the titles of records.

These widget members are easily provided by listing them as Children on
the implementing widget's Template.

[endlessm/eos-sdk#3282]