Port existing cards to Card interface
This takes all the existing card classes, and moves them away from the
Card base class (which is removed). Now they use the new Card interface.
They are also all constructed with widget templates instead of in code.

This allows making some tests simpler; instead of importing the Card
implementations in the tests, we can use a minimal implementation of the
Card interface, which is located in tests/minimalCard.js.

Categories (or, in new terminology, sets) are not yet stored in the
database, so in order to maintain the requirement that all cards are
backed by a record in the database, we create category cards by creating
a ContentObjectModel in code that represents that category.

This also adds a "featured" property to ContentObjectModel to maintain
this functionality while avoiding putting model properties on cards.

[endlessm/eos-sdk#3283]