Remove Card:fade-in property
Fading in is conceptually more of an action than a property; also, the
card's container should determine whether it fades in or not, and ideally
the card's container shouldn't have to be consulted in order to find out
the card's construct properties.

This gets slightly easier if we change the "visible" CSS class to
"invisible" and just don't mess with non-fading-in cards.

[endlessm/eos-sdk#3313]