mediaCard: rework infobox into a card containing the previewer
In our new modular system, we do want infobox and previewer
to be a single card. This reworks the two separate widgets together
so that the infobox container a preview widget as an internal child.

This messes with the lightbox sizing, at is had all this code to
size the previewer and infobox separately.

This commit does not attempt to preserve the lightbox sizing code,
because honestly it was a bit buggy. There was some code in there
to respect a child widget's 'ratio', but that had broken down when
the infobox was added.

The infobox would appear with large empty white spaces when the
caption or attribution was missing, so reworked widgets to hide
when they have no content, as is the norm for cards.
[endlessm/eos-sdk#3317]