Use XML files exactly as Glade generates them
Instead of fighting Glade to uncheck the "visible" property and delete
all the extra junk that it saves to the XML files such as "can_focus"
properties, we use XML files just as they are output by Glade.

This means we make our widget templates visible by default, so we don't
have to call show() on them when we create them. This is counter to what
the convention is in GTK, but it is far more convenient for our code
reviews.

[endlessm/eos-sdk#3313]