kermit: the shard inspection utility
kermit is a program which allows quick inspection of knowledge app
shards. It provides two main functions: dump and grep. Dump will pipe
a particular record's data or metadata for stdout for inspection, while
grep will allow for full regexp search on all content metadata (and in
the case of text records, data as well).

[endlessm/eos-sdk#3241]