reader/presenter: no custom card properties on construction
When we make cards from the factory in the reader presenter, we need
to be careful to only call things guaranteed by the card interface.
Otherwise are cards aren't actually swappable and switching in, say,
a cardA for search results would crash the app.

This shuffles things around, so that page-number is a property of
all cards, and the style variant is inferred from the model's
article number.

[endlessm/eos-sdk#3283]