modules: added HomePageB
This module uses the pack_module functionality in the module
interface to add submodules to its slots, as described
by app.json.

The card_container module has not been created yet, so it is
temporarily replaced by a GtkGrid.

This module also lacks all of the signalling functionality, since
we haven't built the dispatcher yet. As such, it has not been connected
to the general knowledge-lib infrastructure and exists in isolation.

[endlessm/eos-sdk#3339]