Card.set_style_variant_from_model()
For cards which take a style variant, we added a function to the Card
interface which adds the CSS class 'variant0', 'variant1', or 'variant2'
depending on the model's 'article_number' property.

Previously this was done in the cards themselves, but this cuts down on
duplicate code and adds flexibility for adding style variants to other
card types.

The 'snippet[0-2]' and 'reader-card[0-2]' CSS classes are now deprecated
and this is reflected in the style knob parsing code as well.

(Rory & Philip)

[endlessm/eos-sdk#3283]