Move has_margins styling into css theme
Previously we were using a GObject property
on documentCard to determine if the content
within it had margins or not.
It's preferable for all styling info to go
in CSS, and now that we have CSS themes, we
can do that more easily. Keeping the has_margins
property meant that the encyclopedia app,
which will use documentCard, had to know about
it, even thought it is really only a concern
for template A knowledge apps.

Now we have a templateA theme, which includes
the margins it needs on documentCard.

[endlessm/eos-sdk#3370]