Add SearchModule and use it for encyclopedia
This gets rid of the encyclopedia's anomalous HTML search results page,
adding a SearchModule that we can also adapt to other apps.

Lots of files can be removed, such as the HTML, JS, and CSS for the HTML
search results page, and also one of the WebKit plugins that was used to
expose gettext to the search results page.

[endlessm/eos-sdk#3374]