Encyclopedia's SearchModule uses ListArrangement
Instead of creating its own baked-in GtkListBox, the SearchModule in the
Encyclopedia now uses the ListArrangement to display its cards.

It currently creates its Arrangement by creating the named module
"results-arrangement", but that should be changed to be a slot named
"arrangement" when we get submodules and slots on the factory.

[endlessm/eos-sdk#3382]