data/compat: add compatibility gresources for encyclopedia
We are going to start shipping title and background images in the
bundle data itself, as encyclopedia apps become just regular
knowledge apps with a certain selection of modules.

Since our current (soon to be legacy) encyclopedia bundles
don't ship with all that data, we need some legacy encyclopedia
gresources shipping with the library so we continue to support
older encyclopedia bundles in the modular world.
[endlessm/eos-sdk#3384]