Makefile: make gresource pattern rules
More futzing with the gresource dependency rules, which have always
been a pain point.

The original hack we borrowed from gnome-shell didn't work with
built files inside of a gresource, and our current incarnation
was very special cased to our knowledge lib gresource.

This commit attempts to write a bunch a hairy pattern rules that will
handle this for any gresource. You should now be able to something like
bar_DATA = foo.gresource
gresources += foo.gresource
an the pattern rules will take over from there.

It does so by generating a makefile snippet in a .gresource.deps file
which adds files going in the resource as dependencies of the resource
rule, and also to EXTRA_DIST. A bunch of gross string manipulation to
make the .deps file, but hopefully we can for the most part forget about
it now that its in a reusable pattern rule.

We need this cause we are about to add a lot more gresource targets
for our compatibility layer.
[endlessm/eos-sdk#3384]