Load windows through the factory
This doesn't do the work of making the window code any more actually
modular. The encyclopedia apps will still only work with the encyclopedia
window, etc... But this does push a lot more of the app.json parsing
into the factory.

Now the presenter code only touches the json for style knobs (which
need a revamp for the modular system), and the endless switches on
template type in the A/B presenter
[endlessm/eos-sdk#3384]