encyclopedia: use main application class and presenterLoader
We need this as the encyclopedia is becoming just a custom set of
modules, and can't have its own application and loading paths.

Newer encyclopedia apps will have their own gresource and call into
ekn-app-runner. We still ship an eos-encyclopedia bin script for
legacy bundles which will point the runner to our compat
encyclopedia gresources shipped as part of the library.
[endlessm/eos-sdk#3384]