Move custom widgets to widgets directory
We are about to add in the dispatcher, which will broadcast app
level events to whoever cares to listen. Any module can add an
action to the dispatcher for processing.

Widgets that we want to keep reusable, however, should not touch
the dispatcher. The lightbox for example is a perfectly function
gtk widgets that could be used in any other app we need a lightbox
in.

To help make that distinction a bit easier to follow, moved non
module widgets to the widgets directory
[endlessm/eos-sdk#3371]