Create GApplication once in window tests
We don't have to create a new GApplication and run it for each test. We
can do that once per suite in beforeAll. The application is only needed
for creating the window and doesn't actually influence the tests, so test
pollution is unlikely. This makes the tests run a lot faster.

[endlessm/eos-sdk#3396]