All user-facing labels on cards should use_markup
Since the Card interface now markup-escapes strings before sending them
to the UI, all user-facing labels in cards must have their use_markup
property set.

This masks a bigger problem which is that we eventually want our labels
to be able to show rich text, which means we should really store Pango
markup in our databases. But since we've never done that in old bundles,
we can't just start doing it now either. That'll require a version bump
of the database format.

[endlessm/eos-sdk#3406]