presenter: don't double load results when paginating
Not exactly sure why, but the infinite scrolled window is sending
us a lot of 'needs-more-results' queries in quick succession. The
infinite scroll signaling code has always been kinda hairy, so maybe
that broke with new gtk?

Regardless, we need to make sure in the presenter we don't double
load the same "page" in our paginated content.
[endlessm/eos-sdk#3411]