Use item group module in home page B
This changes home page B to use the new item group module. Home page B now
doesn't need to use some of the common functionality from the HomePage
class, and we will eventually remove that class; so this moves all
applicable functionality down to home page A until we are able to replace
that as well.

[endlessm/eos-sdk#3393]