Rework history code
Our history code had a number of bugs and was a frequent breakage point.
This commit does two things to hopefully alleviate the problem.

1 - Remove duplicated state from the presenters, such as "last
origin query" and instead determine the latest query searched by
traversing the history model itself.

2 - Route all UI changes through the history model. In the old world
a button press might trigger a page change, and then on history back
we would trigger that same page change through a different code path.
Now any ui event that would trigger a page change simply alters our
history model, and the actual page is change by the update history
state handler. This keeps all the navigation always happening through
the same path, weather we are navigating to a page for the first time
or through the history. Should make it a lot easier to get the history
code right.
[endlessm/eos-sdk#3160]