Show license of media viewed in lightbox
The lightbox that we use to display media content had always presented the
attribution information (license, owner), but clicking on the button was doing
nothing! Now we can open and display the license file for CC and GNU licenses
when the user presses the button.

Also, when the media displayed in a lightbox is attributed under a license type
that we include, the attribution label changes color on hover.

[endlessm/eos-sdk#3445]