Allow extra props in create_module_for_slot()
We want to be able to create modules from slots at other times than
creation of the parent module. For example, SearchModule should have a
"card_type" slot, and we will call create_module_for_slot() every time we
want to add a card to the search module.

For that to work, we need to be able to specify extra props (such as
"model") when we create the card.

[endlessm/eos-sdk#3433]