Don't create all submodules at construct time
Some slots may be intended for submodules that are created on demand, or
where more than one of the same type of submodule is created. A card-type
slot would be an example of both.

No tests of this functionality directly, but we do change the
HomePageBTemplate tests to assert that the submodules are packed as
descendants of the module.

[endlessm/eos-sdk#3433]