Remove Module.get_submodule()
This method assumed that all submodules were created at construct time.
That's not the case anymore. It looks like we don't really need the method
at all right now, so I removed it. If it would be more useful later, then
we should add it back in but with a different implementation.

[endlessm/eos-sdk#3433]