Add carousel arrangement
This adds an arrangement for document
cards, which show an article's full body
content. It is implemented as a Gtk.Stack
where only one document card is visible at
a time.
Currently, only the reader app uses this
arrangement. This commit refactors its
view to use the arrangement in its main
stack, which retains ownership over
the overview, done, and search results
page.

[endlessm/eos-sdk#3523]