Remove section article page
We are going to move away from having a joined section article page.

Its gonna make template B transitions a little funkier but will make
page management a lot simpler and keep everything modular.

For now, am ripping our the article list of the article page entirely
(it was never really on that page to begin with, just a scrunched up
version of the section page.

Will add it back in a later commit.
[endlessm/eos-sdk#3518]