Added SearchPage
This new page will be used to present search results in the Knowledge Apps.

The purpose of this is to separate, in the future, the behavior of the
SectionPage and SearchPage. What this means for now is a lot of duplicated
code. Nevertheless, this is temporary since these widgets will become
modules in subsequent stories.

[endlessm/eos-sdk#3392]