itemGroup: split into three context modules
Rather than a single itemGroup, we need three separate context modules
for the three separate uses inside the application.

The searchModule should display search results, the itemGroup should
display the contents of a set, and the setGroup should display all
sets.

Added module to the end of all names, as that's the going postfix
for a module ready to be stuck inside a template.

Added a searchABModule which eventually needs to be merged with the
searchModule, however that is written in an encyclopedia specific way
right now. The two will me combined in a later issue.
[endlessm/eos-sdk#3518]