Move private class above main class
This is not technically invalid Javascript, but having the private class
below the main class seems to imply that you can access the private
class everywhere in the file -- which _would_ be invalid Javascript.

Declarations with "const" have block scope, so using the private class
before it was declared is not allowed. It happens to be that we use it
from inside a function which is only run after the const symbol ends up
in the module-level globals, so it works, but it gives the wrong
impression.

Move the private class above the main class; also give it an underscore
to emphasize the fact that it's private, and add a magic "exported"
comment to the top of the file.

[endlessm/eos-sdk#3601]