Use SidebarTemplate for magazine front cover
This requires adding support to the SidebarTemplate for the sidebar width
to be a maximum width rather than a fixed width, and changing the style
knob generator to style the SidebarTemplate's sidebar frame rather than a
widget belonging to the OverviewPage.

This is a transitional state where the OverviewPage packs the
SidebarFrame into itself and packs the submodules into the SidebarFrame.
For that, we need to expose the SidebarFrame's children publicly, which
will later be removed when the OverviewPage is removed entirely.

[endlessm/eos-sdk#3601]