Add OverflowArrangement and use on front cover
This adds an OverflowArrangement (which is a thin wrapper around
SpaceContainer as e.g. ListArrangement around GtkListBox) and uses it
(and therefore the ItemGroupModule) for the reader's magazine front
cover.

This makes the OverviewPage even lighter (on its way to removal) and
requires us to start using the dispatcher for the front cover page.

[endlessm/eos-sdk#3601]