Expose max-children-per-line property on GridArrangement
The grid arrangement module is implemented internally
with a GtkFlowBox. Flow boxes have a property on them
called max-children-per-line, which determines the amount
of space the flow box will request to displaying its
children. We need that property to be publicly available
on the grid arrangement module itself so we can control
how many cards to show per line. So this commit
adds that proxy property and just passes it down to
the private flow box widget.

[endlessm/eos-sdk#3515]