Add HomePageATemplate (HamburgerBasementTemplate)
This adds a template for the home page of template
A apps (but could be used elsewhere). This template
is implemented as a stack with an 'upper' section
and a 'basement' section. The upper section holds
three slots, and the basement section holds one slot.
A tab button (which is part of the template) allows
the user to transition between the upper section
and the basement section.

This obviates the need for a separate categories page,
which is consequently removed in this commit.

[endlessm/eos-sdk#3515]