Add Paper Template
Adds a template that has one slot. The module to
be put in that slot will show up on top of a
'papery' background.

Note that the module in this slot can be swapped
out dynamically at runtime. This is a temporary
measure to support changing the document card
each time a new article model is loaded. Eventually
this may be replaced with some sort of document
card arrangement.

[endlessm/eos-sdk#3557]