Remove OverviewPage
This can now be replaced entirely by modules. Remove the overview page,
its tests, any references to it in the documentation, and remove its
style class in favour of a .sidebar-template style class.

I noticed we were still styling a scrollbar on the overview page which
doesn't exist anymore, so remove that from the CSS file.

The one piece of functionality missing is the overview page's
responsivity to widths of >1366 px. That will be added back in the next
commit.

[endlessm/eos-sdk#3601]