searchBox: dispatch search signals
We also handle the autocomplete queries internally in the module.
I'm not sure if that's the direction we want to go in the long term,
but for now if makes the presenter code much smaller and easier to
convert into a interaction. I think its the right move for now.
[endlessm/eos-sdk#3607]