Use templates for search and article page
The encyclopedia app is getting templatized!
This commit adds an encyclopedia search template
which is used for both the search results page
and the article page. As part of this, the search
banner gets decoupled from the search module
since it is a module in its own right.

The bottom slot of the template does not get filled
on init since, for the article page, its bottom
slot (namely the document card) does not have
a model yet at init time.

This kills the content page since now we have article
and search results pages.

[endlessm/eos-sdk#3557]