articleStackModule: modularize A/B article page
And use the new SHOW_ARTICLE dispatcher signal.

We could try to make this some sort of arrangement, but it does
so much with the asynchronous loading of article pages that I
think its better just left as its own thing.

Also we could add this to some sort of "SimpleTemplate" that just
had one single slot, but that seemed needless.
[endlessm/eos-sdk#3528]