setBannerModule: use SHOW_SET dispatcher action
Before we were using the SET_SELECTED action which is generated by
by a UI click. That actually let to a bug when using the history
module to navigate back to a set, because the click action wouldn't
happen and the proper dispatcher action would not get fired.

Think it makes more sense for the presenter to fire a separate
action when we show a set, rather than having modules listen to
other UI actions to update appearance.
[endlessm/eos-sdk#3528]