WIP - Move action responses to HistoryPresenter
This is a proof of concept to move some duplicate code from the
presenters in to the history presenter. Some responses to dispatcher
actions seem to be simply delegating to the history presenter, so it
seems more natural to listen to the actions from there.

This could go even farther than it does now: we could listen to the
SEARCH_TEXT_ENTERED action in the history presenter as well, but we
would have to move the query sanitization somewhere else.