Add SEARCH_STARTED action
This reinstates the behaviour removed in endlessm/eos-sdk#3557 where the
search banner displays a message indicating that the search is in
progress. This message is currently not visible on the template A and B
apps. It should be visible in the Encyclopedia app once the Encyclopedia
presenter starts to use dispatcher actions for its search.

For this, we add a SEARCH_STARTED dispatcher action that is dispatched
right after the user activates a search and the query is sent to the
search engine.

[endlessm/eos-sdk#3559]