Add SEARCH_FAILED action
In order to unify our search experience across different knowledge apps,
our search module should display an error page like the Encyclopedia app
has always done.

This adds the appropriate dispatcher action and has the knowledge app
presenter dispatch it, but doesn't actually work in the app yet; I don't
want to implement it in the SearchABModule because I'm later removing
that module on this branch.

[endlessm/eos-sdk#3559]