Dispatchify the encyclopedia search
Previously we had start_search(), finish_search(), and
finish_search_with_error() methods on the search module. Instead of that,
the Encyclopedia presenter now dispatches actions, and the search module
listens to them.

This allows getting rid of a lot of the temporary properties that granted
the ability for modules to reach inside other modules, which makes the
code quite a lot cleaner.

[endlessm/eos-sdk#3559]