Remove hardcoding of search query highlight
Previously, the search query in 'Search results for "(query)"' was
colored differently and had a different weight than the rest of the
text. This only applied to the Encyclopedia app.

After merging the SearchABModule into the SearchModule, all apps
started showing that highlight. This once again confines that highlight
to the Encyclopedia app and additionally makes it customizable via CSS.

[endlessm/eos-sdk#3559]