Put no-results message and error message together
In the search module, we had two almost identical labels for the no-
results message and the error message. Just use one label for both and
change its style class depending on the context that it's used for.

Add the same style classes to the label in the search banner module so
that we can style them both in parallel.

[endlessm/eos-sdk#3559]