encyclopedia/presenter: use dispatcher for article page
We can use the same articleStackModule we use for the article page
on template A/B, just without any animations

This gets rid of the fullscreen code for the webview, but it was already
broken as it tried to set a Gtk.Alignment property rather than working
with the new paper template. We already have sdk#3623 to fix this, so I
think we can just drop the broken code for now.
[endlessm/eos-sdk#3647]