inArticleSearch: revamp theme
When we moved the in article search inside the documentCard we fixed
its allocation, before it was outside of the paper template and
far too big. However moving it inside the paper template also means
it now shows up on the paper background.

We could try to undo the paper template theme so it doesn't give a
background, and put it on all the other widgets, but that would be
a bit ticky.

Instead decided to retheme the search widget so it worked on the
new background. The widget had some serious problems anyway. No
hover state on buttons, next button before previous button (which
doesn't match our browser or anything else).

Fixed those up and tried to roughly match the article them, so think
we should be a better place till this gets real design attention.
[endlessm/eos-sdk#3647]