Don't lock_ui and unlock_ui
Now that we update the UI to indicate that a search is in progress, it's
not necessary to render the whole UI insensitive while the search is
going on. And now that we have the dispatcher, it's quite easy to
subscribe the window modules to the correct actions. In order to maintain
all the functionality we already had, we also add a SET_READY action
that's dispatched when a set has filled up with its models.

Instead of lock_ui() and unlock_ui() we now have a private window method
which changes the mouse cursor but doesn't set everything insensitive.

There is a bunch of duplicate code in Window and ReaderWindow now, but
that will be removed soon when we merge them.

[endlessm/eos-sdk#1904]