searchProvider: check app-id when mapping from object path
Because application ids can contain dashes and underscores but dbus
object paths can only contain underscore, we actually lose
information when translating from an app id to a object path.

We need to check if the last underscore in our app id name is in
fact a valid underscore in the locale part of an knowledge app id.

We should also check if there's an actually app with that id at
all before making a search provider for it.
[endlessm/eos-sdk#3655]