paperTemplate: restore old sizing behavior
This will mimic the old gtk alignment, trying to give at least 700
pixels to the paper frame and taking up 40% of the extra space after
that.

With our new conversations with design about resizing, might have
something smarter to do here soon. But in the meantime we can use
the old functionality, which performs better at small resolutions
than the current implementation.
[endlessm/eos-sdk#3665]