scss: line-height uses unitless numbers
Unitless number avoid a lot of the unexpected inheritance problems
with line heights with units.

https://developer.mozilla.org/en-US/docs/Web/CSS/line-height

With this change our line heights will scale with font size, so as
long as we never choose a line height > 1 there should be no way to
see text overlap like we currently do with a high text scale factor.

One spot where the styling will be slightly different is the
wikimedia infoboxes. We were formally trying to apply line height to
a table, which apparently does nothing, things only work if you apply
the line height to the tr, th and tds in the table.

So the old code was accidentally setting all line heights to 22px
in the infobox. That's bad, as things do appear at different font
sizes in the infobox. But changing things to properly scale with
font size creates some slight visual differences on some infoboxes,
as the font size for headers in the infobox is not actually static
across all wikimedia articles.
[endlessm/eos-sdk#3666]