knowledgeDocumentCard: drop min content view width
We bumped it all the way to 700 for the paper template of the
encyclopedia app, but now that the paper size is handled elsewhere
we can drop it back to the old value of 400.

The 700 min content width was actually breaking things for certain
sizes of template A app, as the uncollapsed table of contents is
>> 100 pixels wide, and we don't switch to collapsed state until
we hit 800 pixels in width. The end result is that sometime the
webview would get less then its minimum allocated and would behave
poorly.
[endlessm/eos-sdk#3667]