Add SetObjectModel with child_tags prop
To correspond to the EKN vocabulary added in the companion pull request
on eos-knowledge-db-build, we add a SetObjectModel to our family of
ContentObjectModels. The SetObjectModel has a child_tags property to
indicate the tags with which articles must be tagged if they are members
of this set.

This changes the v1 compatibility code to add SetObjectModels as runtime
objects to the database. Since the whole concept didn't exist in v1
databases, there is no backwards compatibility worry there. In v2
app.jsons, the SetObjectModels are expected to be in the database.

[endlessm/eos-sdk#3570]