Dispatch FIRST_LAUNCH when app should show itself
Moving one more view action out of the presenter; FIRST_LAUNCH is
dispatched when the window should present itself. This is a bit
complicated for the reader app since it only presents itself once the
content is loaded, and that means we have to store a bit of extra state
in order to remember how we were launched, unfortunately.

This will be necessary for the brand screen.

[endlessm/eos-sdk#3678]