Add "brand screen" to startup of mesh interaction
The mesh interaction now has an optional "brand-screen" submodule, which
can be any widget module. When starting an app with a mesh interaction,
if the brand screen is present, it is shown instead of the home page.
Then, when a certain time has passed (currently 1.5 seconds), the brand
screen fades out (duration 0.5 seconds) to the home page.

[endlessm/eos-sdk#3678]