Add suggested items logic to reader app
Populate the new suggested items modules with
content in the reader app. For now, we just
provide them with random articles.

One trick that needed taking care of is that
although the suggested articles need to be
random for each search, they should be tied
to a history item. So that if you go back
to a search results page with that had no
results for your query, you see the same
'random articles' each time. The way we handle
this is by using the query itself as a seed
for random number generation. So although the
results are pseudorandom for the user, the same query
will always generate the same random suggested results.

[endlessm/eos-sdk#3675]