compat: use anonymous modules for arrangements
Just doing this for arrangements that are only used in one place
for the time being.

Not going hog wild with anonymous modules for the time being. I plan
to split each templates compat json into its own file in the future
and that seems like a better time to shift everything around.
[endlessm/eos-sdk#3712]