engine: add get_default to import object
We were inconsistent on where we added the method between engine
and dispatcher--one of the import object the other on the class
object.

I prefer the former as its a little less verbose. The latter has
a closer parallel with get_default methods on C glib objects, but
if we ever do change engine to a C class we'd be changing all its
imports and namespacing anyway.
[endlessm/eos-sdk#3713]