homePageATemplate: give default values for button labels
We'll have the home page template buttons default to SEE ALL CATEGORIES
and HOME so we don't have to set it in the app.json.

This is a deliberately short sighted fix, as eventually we will want
to just the class as a general purpose template in other use cases.
But for now we don't want to handle localization in app.json files.
[endlessm/eos-sdk#3681]