Highlights module
This adds a HighlightsModule which implements most of the home page of
the new travel app. It has an arrangement for any cards at the top,
followed by arrangements for each of the non-featured sets ("themes")
from the database.

Each theme is headed by a card which can also be clicked.

It's designed to be used in a situation where all the sets and cards are
dispatched when the app starts up. For this we will add a separate
interaction in the next commit.

InVision mockup: https://endlessm.invisionapp.com/d/main#/console/4678196/107219771/preview
InVision wireframe: https://endlessm.invisionapp.com/d/main#/console/4678196/106249329/preview

[endlessm/eos-sdk#3699]