Buffet interaction
We need a separate interaction because the travel app requires populating
itself with all the app's sets and articles when the app starts up. Let's
call this the "buffet" because it sets out everything it has and lets the
user pick what they want.

This includes some rudimentary CSS styling to make the new modules at
least not hurt one's eyes.

[endlessm/eos-sdk#3699]