Added CSS Card sizing classes
The cards in the new modular system will behave in a more responsive way. Here
we add the CSS classes that handle the sizing of cards in this new system.

These CSS classes are added/removed to the card widgets according to the
dimensions of the cards. This is done with the `update_card_sizing_classes`
method in the `Card` interface, which should be invoked in the implementations
of the `vfunc_size_allocate` method.

[endlessm/eos-sdk#3720]