Make fixed-size cards really fixed-size
This was a difficult one to track down. A test in reader card started
failing locally for me. Bisecting showed that it started failing when we
renamed the reader presenter to aisle interaction so that its test ran
before the reader card test. The reader card test failed because we added
CSS to the default screen in aisle interaction, causing the fonts to get
bigger and the wrapped text to be larger than 200 pixels, exposing a bug
where only the minimal width and height were set to the fixed values, and
the natural width and height were still free.

So there really was a bug, and we fix it by overriding the natural width
of the fixed-size cards.

[endlessm/eos-sdk#3700]