Add history and set page to buffet interaction
We are going to add the thematic module in this branch, so we need to add
the necessary functionality to the buffet interaction to support it. The
buffet interaction gains a history presenter and can switch back and
forth between the home page and the set page.

It switches to the set page on the SET_CLICKED action.

[endlessm/eos-sdk#3700]