Added SquareGuysArrangement
An arrangement implementation that keeps the square proportions of its cards.

Two rows of cards are displayed, even if more widgets are added. Eight cards are
displayed (4x2) when the arrangement width is greater than 1200px. Six cards are
displayed (3x2) when the width is between [600px, 1200px). The minimum width for
the arrangement is 600px.

The height of the arrangement is constrained to 600px or 400px, depending on the
width.

[endlessm/eos-sdk#3696]