Add the ability to filter suggested categories
This adds a featured-only flag to SuggestedCategoriesModule. If the flag
is set, then we only add featured sets to the arrangement.

This would be more properly implemented as a filter for the arrangement,
but filters and sorting functions are one of the few things we haven't
implemented for the modular system yet.

[endlessm/eos-sdk#3698]