Context banner
This adds a module that displays a different heading depending on what
action was last dispatched. When the home page is shown, it displays
"Highlights" (internationalized). When a set is shown, it displays the
set's title. When search results are shown, it displays the number of
results and the search text.

[endlessm/eos-sdk#3698]