ekn-util: remove unused global uri scheme workaround
We had problem registering a uri scheme on the default webkit
context through javascript because on an atexit handler in webkit.
See https://bugzilla.gnome.org/show_bug.cgi?id=729611

Now that we have a context per webview this is no longer a problem
and we should not need the introspection hack anymore.
[endlessm/eos-sdk#3670]