CSS styling
Various CSS styling fixes that I made for the skeleton app demo. This is
basically just a commit of all the CSS changes I have up to now.

I decided to use px instead of em for this app because we have designs
that take font scaling into account, so we don't want to use our usual
EosWindow font scaling.

[endlessm/eos-sdk#3758]