slidingPanelOverlay: style property for fill percentage
Adding a custom style property so we can keep this to css. It's
on the slidingPanel, but only for when it is used with
slidingPanelOverlay.

This gives us better resizing behavior than the previous 200 pixel
margin on either side of the previous next cards. Stole the exact
value from the css in the html prototype.
[endlessm/eos-sdk#3719]