Display top search box only if page has none
In our existing knowledge apps, there is always a search box on the home
page, so we don't display the search box in the title bar when the home
page is showing. However, this was hardcoded, and we need it to change
for our new travel app.

Instead, we detect if there's a SearchBox module somewhere on the page
being displayed. If there is, we hide the title bar's search box.

[endlessm/eos-sdk#3758]