Accumulate all models before displaying them on screen
Instead of adding cards to the screen as soon as
we receive them, wait for all models to be fetched
and then add them all at the same time. This
prevents the weird flash where you see some
cards on screen and then they get replaced
by others after a second or so.

[endlessm/eos-sdk#3798]