meshInteraction: consolidate history item handling
Encyclopedia and template A/B really don't need their own switch
statements. They are the same except for handling animations when
switching in a new article page.

Plus with the diverged switches, we had introduced a least one bug.
Encyclopedia would not autofill its search entry when navigating the
history model to a search page.
[endlessm/eos-sdk#3561]