searchBox: only grab focus if mapped
Previously we would grab focus on any search box that was visible.
On encyclopedia this would grab focus of the search bar on the
second page when opening the app, which was visible but not mapped.

Typing without first clicking on the search box would spew error
to the terminal, as you were typing on a widget that was not on
screen
[endlessm/eos-sdk#3561]