Shouldn't go back to home if we didn't start there
Previously when launching an app via a search or a search result, the
history back button was active and you could go back to the home page,
even though you had never visited there. This changes things so that the
home page item is only added to the history stack if you launched the app
such that it started on the home page.

It also moves the focusing of the search bar to the same launcher handler
instead of the mesh interaction's constructor, and fixes a small bug
where it was assumed that there would always be a previous item in the
history.

[endlessm/eos-sdk#3795]